transparencia/requirements.txt
django==1.11.23
mysqlclient
uwsgi
celery
redis
flake8
django-environ
django-bootstrap4
django-easy-pdf
xhtml2pdf>=0.2b1
requests
asn1crypto
bcrypt
cffi
cryptography
django-bootstrap4
django-colorful
django-icons
django-tinymce
enum34
idna
ipaddress
paramiko
pyasn1
pycparser
PyNaCl
pysftp
pytz
six
environ


